pyproject.toml
[project]
name = "ai-app-developer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "longchain>=0.4.2",
    "pip>=26.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
